# Feature/Optimize Node List Serialization [PLAT-981]