--- a/api/base/serializers.py
+++ b/api/base/serializers.py
@@ -140,7 +140,8 @@ class ShowIfAdminScopeOrAnonymous(ConditionalField):
 
     def should_show(self, instance):
         request = self.context.get('request')
-        return request and (request.user.is_anonymous or utils.has_admin_scope(request))
+        has_admin_scope = instance.has_admin_scope if hasattr(instance, 'has_admin_scope') else utils.has_admin_scope(request)
+        return request and (request.user.is_anonymous or has_admin_scope)
 
 
 class HideIfRegistration(ConditionalField):
@@ -203,7 +204,8 @@ class HideIfWikiDisabled(ConditionalField):
 
     def should_hide(self, instance):
         request = self.context.get('request')
-        return not utils.is_deprecated(request.version, min_version='2.8') and not instance.has_addon('wiki')
+        has_wiki_addon = instance.has_wiki_addon if hasattr(instance, 'has_wiki_addon') else instance.has_addon('wiki')
+        return not utils.is_deprecated(request.version, min_version='2.8') and not has_wiki_addon
 
 
 class HideIfNotNodePointerLog(ConditionalField):