… osf/models/node.py