<!-- Before submit your Pull Request, make sure you picked
     the right branch:

     - For hotfixes, select "master" as the target branch
     - For new features, select "develop" as the target branch
     - For release feature fixes, select the relevant release branch (release/X.Y.Z) as the target branch -->

## Purpose

Serializing lists of nodes are slow - causing issues with collections endpoint on MyProjects page (which embeds LinkedNode), and issues with UserNodes.

## Changes

Reducing n+1 queries on the following locations on the NodeSerializer:
- tags field
- ~node_license field~ **
- analytics_key field
- current_user_can_comment field
- current_user_permissions field
- ~license~ **
- wikis relationship
- parent relationship
- root relationship
- subjects field

** n+1 query problem still exists on licenses (because we have to fetch license recursively).  However, number of license queries has been halved. License is fetched once per node and put on serializer context for the other license field to use, credit @sloria 

Improvements made by annotating node queryset or using `prefetch_related`. Consolidated in `NodeOptimizationMixin` so any queryset of nodes can use to reduce # of requests.

## Improvements
###  CollectionsList, fetching 5 collections, with 10 nodes apiece 
`http://localhost:8000/v2/collections/?related_counts=linked_registrations%2Clinked_nodes&page[size]=5&sort=date_created&embed=linked_nodes`

Before: 538 queries in 2149.25 ms
After: 106 queries in 369.20 ms

### NodeList, fetching 10 nodes
`http://localhost:8000/v2/nodes/`

Before: 109 queries in 332.35 ms
After: 23 queries in 100.92 ms

### UserNodes, fetching 10 nodes
`https://localhost:8000/v2/users/me/nodes/`

Before: 88 queries in 231.22 ms
After: 20 queries in 85.98 ms


## QA Notes

<!-- Does this change need QA? If so, this section is required.
     - Is cross-browser testing required/recommended?
     - Is API testing required/recommended?
     - What pages on the OSF should be tested?
     - What edge cases should QA be aware of?
-->

## Documentation

None needed.

## Side Effects

<!-- Any possible side effects? -->

## Ticket
https://openscience.atlassian.net/browse/PLAT-981